Added doc to .gitignore to avoid accidental add
Some versioned files were moved to the doc directory in previous
commit. This directory was used to save generated HTML documentation
and was formely ignored by git. It became not ignored, but this creates
risk that after a pull the multitude of files may be accidentally added
for versioning. Reverting the doc entry in .gitignore until everyone
wipes the doc directory.